+++ src/restfw/testing/resource_testing.py
@@ -29,7 +29,7 @@ class RequestsTester(object):
         self.calls_count = 0
 
     def __call__(self, params=DEFAULT, headers=None, result=None, result_headers=None,
-                 exception=None, status=None):
+                 exception=None, status=None, description=None):
         """
         :type params: dict or list or str or None
         :type headers: dict or None
@@ -37,6 +37,7 @@ class RequestsTester(object):
         :type result_headers: dict or None
         :type exception: HTTPException
         :type status: int
+        :type description: str or None
         """
         raise NotImplementedError
 
@@ -45,7 +46,7 @@ class RequestsTester(object):
 class GetRequestsTester(RequestsTester):
 
     def __call__(self, params=DEFAULT, headers=None, result=None, result_headers=None,
-                 exception=None, status=None):
+                 exception=None, status=None, description=None):
         """
         :type params: dict or list or str or None
         :type headers: dict or None
@@ -53,6 +54,7 @@ class GetRequestsTester(RequestsTester):
         :type result_headers: dict or None
         :type exception: HTTPException
         :type status: int
+        :type description: str or None
         """
         self.calls_count += 1
         params = params if params is not DEFAULT else {}
@@ -79,7 +81,7 @@ class GetRequestsTester(RequestsTester):
 class PutRequestsTester(RequestsTester):
 
     def __call__(self, params=DEFAULT, headers=None, result=None, result_headers=None,
-                 exception=None, status=None):
+                 exception=None, status=None, description=None):
         """
         :type params: dict or list or str or None
         :type headers: dict or None
@@ -87,6 +89,7 @@ class PutRequestsTester(RequestsTester):
         :type result_headers: dict or None
         :type exception: HTTPException
         :type status: int
+        :type description: str or None
         """
         self.calls_count += 1
         params = params if params is not DEFAULT else {}
@@ -110,7 +113,7 @@ class PutRequestsTester(RequestsTester):
 class PatchRequestsTester(RequestsTester):
 
     def __call__(self, params=DEFAULT, headers=None, result=None, result_headers=None,
-                 exception=None, status=None):
+                 exception=None, status=None, description=None):
         """
         :type params: dict or list or str or None
         :type headers: dict or None
@@ -118,6 +121,7 @@ class PatchRequestsTester(RequestsTester):
         :type result_headers: dict or None
         :type exception: HTTPException
         :type status: int
+        :type description: str or None
         """
         self.calls_count += 1
         params = params if params is not DEFAULT else {}
@@ -141,7 +145,7 @@ class PatchRequestsTester(RequestsTester):
 class PostRequestsTester(RequestsTester):
 
     def __call__(self, params=DEFAULT, headers=None, result=None, result_headers=None,
-                 exception=None, status=None):
+                 exception=None, status=None, description=None):
         """
         :type params: dict or list or str or None
         :type headers: dict or None
@@ -149,6 +153,7 @@ class PostRequestsTester(RequestsTester):
         :type result_headers: dict or None
         :type exception: HTTPException
         :type status: int
+        :type description: str or None
         """
         self.calls_count += 1
         params = params if params is not DEFAULT else {}
@@ -172,7 +177,7 @@ class PostRequestsTester(RequestsTester):
 class DeleteRequestsTester(RequestsTester):
 
     def __call__(self, params=DEFAULT, headers=None, result=None, result_headers=None,
-                 exception=None, status=None):
+                 exception=None, status=None, description=None):
         """
         :type params: dict or list or str or None
         :type headers: dict or None
@@ -180,6 +185,7 @@ class DeleteRequestsTester(RequestsTester):
         :type result_headers: dict or None
         :type exception: HTTPException
         :type status: int
+        :type description: str or None
         """
         self.calls_count += 1
         params = params if params is not DEFAULT else {}